:recycle: コンポーネント修正
