+++ apps/frontend/src/components/Card/card.ts
@@ -1,0 +1,18 @@
+type Article = {
+  id: string;
+  title: string;
+  body: string;
+  imageUrl: string;
+  categoryName: string;
+  href: string;
+  author: {
+    name: string;
+    href: string;
+    imageUrl: string;
+  };
+  datetime: string;
+  date: string;
+  readingTime: string;
+};
+
+export default Article;
