+++ apps/frontend/src/components/Footer/index.tsx
@@ -3,7 +3,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import React from 'react';
 
 const iconStyle: React.CSSProperties = { fontSize: 25 };
-const Footer = () => (
+const Index = () => (
   <>
     <div className='bg-white pt-4 sm:pt-10 lg:pt-12'>
       <footer className='max-w-screen-2xl px-4 md:px-8 mx-auto'>
@@ -11,31 +11,31 @@ const Footer = () => (
           {/* nav - start */}
           <nav className='flex flex-wrap justify-center md:justify-start gap-x-4 gap-y-2 md:gap-6 mb-8'>
             <a
-              href='#'
+              href='apps/frontend/src/components#'
               className='text-gray-500 hover:text-indigo-500 active:text-indigo-600 transition duration-100'
             >
               About
             </a>
             <a
-              href='#'
+              href='apps/frontend/src/components#'
               className='text-gray-500 hover:text-indigo-500 active:text-indigo-600 transition duration-100'
             >
               Investor Relations
             </a>
             <a
-              href='#'
+              href='apps/frontend/src/components#'
               className='text-gray-500 hover:text-indigo-500 active:text-indigo-600 transition duration-100'
             >
               Jobs
             </a>
             <a
-              href='#'
+              href='apps/frontend/src/components#'
               className='text-gray-500 hover:text-indigo-500 active:text-indigo-600 transition duration-100'
             >
               Press
             </a>
             <a
-              href='#'
+              href='apps/frontend/src/components#'
               className='text-gray-500 hover:text-indigo-500 active:text-indigo-600 transition duration-100'
             >
               Blog
@@ -46,7 +46,7 @@ const Footer = () => (
           {/* social - start */}
           <div className='flex gap-8'>
             <a
-              href='#'
+              href='apps/frontend/src/components#'
               target='_blank'
               className='text-gray-400 hover:text-gray-500 active:text-gray-600 transition duration-100'
             >
@@ -57,7 +57,7 @@ const Footer = () => (
             </a>
 
             <a
-              href='#'
+              href='apps/frontend/src/components#'
               target='_blank'
               className='text-gray-400 hover:text-gray-500 active:text-gray-600 transition duration-100'
             >
@@ -68,7 +68,7 @@ const Footer = () => (
             </a>
 
             <a
-              href='#'
+              href='apps/frontend/src/components#'
               target='_blank'
               className='text-gray-400 hover:text-gray-500 active:text-gray-600 transition duration-100'
             >
@@ -89,4 +89,4 @@ const Footer = () => (
   </>
 );
 
-export default Footer;
+export default Index;
