+++ apps/frontend/src/pages/_app.tsx
@@ -1,8 +1,8 @@
 import '@/styles/globals.css';
 import { ApolloProvider } from '@apollo/client';
 
-import Footer from '@/components/footer';
-import Header from '@/components/header';
+import Footer from '@/components/Footer/index';
+import Index from '@/components/Header';
 import { client } from '@/lib/apollo';
 
 import type { AppProps } from 'next/app';
@@ -11,7 +11,7 @@ function MyApp({ Component, pageProps }: AppProps<{}>): JSX.Element {
   return (
     <ApolloProvider client={client}>
       <div className='max-w-screen-2xl'>
-        <Header />
+        <Index />
 
         <Component {...pageProps} />
